middleware/requirements.txt
fastapi==0.75.1
redis==4.2.2
uvicorn==0.17.6
pydantic==1.9.0
SQLAlchemy==1.4.35
psycopg2-binary==2.9.3
requests==2.27.1
kafka-python==2.0.2